Remove all redundant suffixes from filenames
