+++ src/frame/js/reducers/integrations.js
@@ -1,7 +1,7 @@
-import { RESET_APP } from '../actions/app-actions';
-import { RESET } from '../actions/common-actions';
+import { RESET_APP } from '../actions/app';
+import { RESET } from '../actions/common';
 
-import { SET_ERROR, UNSET_ERROR, SET_WECHAT_QR_CODE, SET_TWILIO_INTEGRATION_STATE, RESET_TWILIO_INTEGRATION_STATE, SET_MESSAGEBIRD_INTEGRATION_STATE, RESET_MESSAGEBIRD_INTEGRATION_STATE, RESET_INTEGRATIONS, SET_VIBER_QR_CODE, SET_TRANSFER_REQUEST_CODE, RESET_TRANSFER_REQUEST_CODE } from '../actions/integrations-actions';
+import { SET_ERROR, UNSET_ERROR, SET_WECHAT_QR_CODE, SET_TWILIO_INTEGRATION_STATE, RESET_TWILIO_INTEGRATION_STATE, SET_MESSAGEBIRD_INTEGRATION_STATE, RESET_MESSAGEBIRD_INTEGRATION_STATE, RESET_INTEGRATIONS, SET_VIBER_QR_CODE, SET_TRANSFER_REQUEST_CODE, RESET_TRANSFER_REQUEST_CODE } from '../actions/integrations';
 
 const INITIAL_STATE = {
     messenger: {
