+++ src/frame/js/reducers/user.js
@@ -1,5 +1,5 @@
-import { SET_USER, RESET_USER, UPDATE_USER } from '../actions/user-actions';
-import { RESET } from '../actions/common-actions';
+import { SET_USER, RESET_USER, UPDATE_USER } from '../actions/user';
+import { RESET } from '../actions/common';
 
 const INITIAL_STATE = {};
