+++ src/frame/js/store/configure-dev.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore, compose } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 
-import { RootReducer } from '../reducers/root-reducer';
+import { RootReducer } from '../reducers/root';
 
 const finalCreateStore = compose(
     applyMiddleware(thunkMiddleware),
@@ -12,7 +12,7 @@ export function configureStore(initialState) {
     const store = finalCreateStore(RootReducer, initialState);
 
     if (module.hot) {
-        module.hot.accept('../reducers/root-reducer', () => store.replaceReducer(require('../reducers/root-reducer').RootReducer));
+        module.hot.accept('../reducers/root', () => store.replaceReducer(require('../reducers/root').RootReducer));
     }
 
     return store;
