+++ src/frame/js/store/configure-prod.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 
-import { RootReducer } from '../reducers/root-reducer';
+import { RootReducer } from '../reducers/root';
 
 const createStoreWithMiddleware = applyMiddleware(
     thunkMiddleware,
