+++ src/frame/stylesheets/reply-actions.less
@@ -1,4 +1,4 @@
-.sk-reply-actions-container {
+.sk-reply-container {
     text-align: right;
     position: relative;
     display: block;
