+++ test/specs/reducers/integrations.spec.js
@@ -1,6 +1,6 @@
-import { IntegrationsReducer } from '../../../src/frame/js/reducers/integrations-reducer';
+import { IntegrationsReducer } from '../../../src/frame/js/reducers/integrations';
 
-import { SET_WECHAT_QR_CODE, SET_ERROR, UNSET_ERROR, RESET_INTEGRATIONS, SET_TWILIO_INTEGRATION_STATE, RESET_TWILIO_INTEGRATION_STATE } from '../../../src/frame/js/actions/integrations-actions';
+import { SET_WECHAT_QR_CODE, SET_ERROR, UNSET_ERROR, RESET_INTEGRATIONS, SET_TWILIO_INTEGRATION_STATE, RESET_TWILIO_INTEGRATION_STATE } from '../../../src/frame/js/actions/integrations';
 
 describe('Integrations reducer', () => {
